Anonymize a vector.

anonymize anonymizes a vector .x by first salting it with salt and then hashing it with hash. See both functions for additional documentation.

anonymize(.x, .algo = "sha256", .seed = 0, .chars = letters,
  .n_chars = 5L, ...)

Arguments

.x
a vector.

.algo
the name of the algorithm.

.seed
an integer to seed the random number generator.

.chars
set of characters to salt with.

.n_chars
an integer; number of characters to salt with.

...
additional arguments to be passed to hash.

Value

An anonymized version of the vector.

Details

The user is advised to check out Wikipedia for more information.

Examples

set.seed(1)
anonymize(letters)

#>  [1] "88004902989219b709ab0b6df47b4bd5a16d77b54b100bf90cd1892f6209eb69"
#>  [2] "18246e832ba9ebb0f66e00234ce9bd13072b2320197d5234bd2f9923ba7266c7"
#>  [3] "7e34242b3692ea9869e59789304661f0a0b57f95c2171fd5dda957cefc174ebd"
#>  [4] "53022e73ef8c8e6c4b9c28fe88b2f7c3dc9752614976896d26a6a5bac262696c"
#>  [5] "47af26e7c1c28b2bd1885aaade5ea329468c9722748ead0f4ad37038aa5431d8"
#>  [6] "ca49b69a827d7f354c02463a7a591b64400ad438a21d717389a31255844b0d9c"
#>  [7] "e1d482341c8da0f7dcb148ff25ded4ece1e950ef23ad4f7e43c8add47710302e"
#>  [8] "11472a257f5840e1a612b9cbcedf3925e8dbb5ef91be991eafd438203c0ac1f0"
#>  [9] "f0d22c7f3ef30628d3c7148f6c7e778b4a828229b1ab28535693d3061f8fc656"
#> [10] "794801e15d2796dd4662d97442c87bebd8264dde055312451d4c2b6978f786b1"
#> [11] "c4c74732547eb74de5d6dd4ff5c7cb125d4ba94cf832c02e57822d6d3a5571c9"
#> [12] "ea9987e5bbca0c5d709983d2fc20332b3ed55637d6a87909628b3316a3d583ff"
#> [13] "760c7a24f910f33231a3dd5a5ae0b7c44a2047ccc7212afe5692f8582361c344"
#> [14] "f18eba4da2043b26decc2a9009a0d46956dc6098c16028e3578a759b31359146"
#> [15] "0d00e7489fe0bd2600b0e0bf5b02d69c651d94de68bdb18e6a063679d99c3233"
#> [16] "5eaaf1c642f45cf2e73d4fb6975055b8d14d00f1f79450bec2e3cdc493cd65bb"
#> [17] "9f8a4a8a6252634125028830028aba714a52d78a198dca90fd911e646595b021"
#> [18] "ba59071827c458be24a3650e8691cdb04a9f17fa9208466cc5656e1e9e95f5e6"
#> [19] "e6338111f03e2df2ddf416c91d321d7502cb8e64afa1a3f564cc57d31d505ffd"
#> [20] "162599bb4d510f4e80302d05df25d223ac6e56d8d9fcdeb7ab7a0d9332f6c083"
#> [21] "2a6273277d80c9f93c9728b0d468f28e4453d764a6cff0a2a33c2f3e8fba03d8"
#> [22] "946027b956078f2e3a73b90342c693c055e8e780f917d8d0adf0cc57253fd7d4"
#> [23] "4c949369d988fa516e055c100acf70138dacf1ec78a5d7a12e029ed8544e6a50"
#> [24] "a93f41e5832362430a1b7a0d85d16b199271c3eb54672680536332351e749727"
#> [25] "a26345dac5fa2e703e18f69fa48789876456ecca23202aebf6fdc2c6002fc30a"
#> [26] "520ae4c73eb3de37cf4b365de2f0b59bfc6ff8ec81294edf531ebbae178394e3"