Salt a vector.

salt takes a vector .x and returns a salted version of it. The algorithm for salting a vector is:

Select a random sample of characters of length .n_chars from .chars. Call this random sample .y.

Concatenate .y, the vector .x, and .y again in a vectorized fashion.

salt(.x, .seed = NULL, .chars = letters, .n_chars = 5L)

Arguments

.x
a vector.

.seed
an integer to seed the random number generator.

.chars
set of characters to salt with.

.n_chars
an integer; number of characters to salt with.

Value

A salted version of the vector.

Details

The user is advised to check out Wikipedia for more information.

Examples

# Use various number of characters
salt(letters, .n_chars = 0L)

#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z"

salt(letters, .n_chars = 1L)

#>  [1] "gag" "gbg" "gcg" "gdg" "geg" "gfg" "ggg" "ghg" "gig" "gjg" "gkg" "glg"
#> [13] "gmg" "gng" "gog" "gpg" "gqg" "grg" "gsg" "gtg" "gug" "gvg" "gwg" "gxg"
#> [25] "gyg" "gzg"

salt(letters, .n_chars = 5L)

#>  [1] "joxfxajoxfx" "joxfxbjoxfx" "joxfxcjoxfx" "joxfxdjoxfx" "joxfxejoxfx"
#>  [6] "joxfxfjoxfx" "joxfxgjoxfx" "joxfxhjoxfx" "joxfxijoxfx" "joxfxjjoxfx"
#> [11] "joxfxkjoxfx" "joxfxljoxfx" "joxfxmjoxfx" "joxfxnjoxfx" "joxfxojoxfx"
#> [16] "joxfxpjoxfx" "joxfxqjoxfx" "joxfxrjoxfx" "joxfxsjoxfx" "joxfxtjoxfx"
#> [21] "joxfxujoxfx" "joxfxvjoxfx" "joxfxwjoxfx" "joxfxxjoxfx" "joxfxyjoxfx"
#> [26] "joxfxzjoxfx"

salt(letters)

#>  [1] "yrqbfayrqbf" "yrqbfbyrqbf" "yrqbfcyrqbf" "yrqbfdyrqbf" "yrqbfeyrqbf"
#>  [6] "yrqbffyrqbf" "yrqbfgyrqbf" "yrqbfhyrqbf" "yrqbfiyrqbf" "yrqbfjyrqbf"
#> [11] "yrqbfkyrqbf" "yrqbflyrqbf" "yrqbfmyrqbf" "yrqbfnyrqbf" "yrqbfoyrqbf"
#> [16] "yrqbfpyrqbf" "yrqbfqyrqbf" "yrqbfryrqbf" "yrqbfsyrqbf" "yrqbftyrqbf"
#> [21] "yrqbfuyrqbf" "yrqbfvyrqbf" "yrqbfwyrqbf" "yrqbfxyrqbf" "yrqbfyyrqbf"
#> [26] "yrqbfzyrqbf"

# Use other sets of characters to salt with
salt(letters, .chars = letters[1:2])

#>  [1] "ababaaababa" "abababababa" "ababacababa" "ababadababa" "ababaeababa"
#>  [6] "ababafababa" "ababagababa" "ababahababa" "ababaiababa" "ababajababa"
#> [11] "ababakababa" "ababalababa" "ababamababa" "ababanababa" "ababaoababa"
#> [16] "ababapababa" "ababaqababa" "ababarababa" "ababasababa" "ababatababa"
#> [21] "ababauababa" "ababavababa" "ababawababa" "ababaxababa" "ababayababa"
#> [26] "ababazababa"