Unsalt a vector.

salt takes a vector .x and returns an unsalted version of it. The algorithm for unsalting a vector is:

Select a random sample of characters of length .n_chars from .chars. Call this random sample .y.

Substitute .y out of the vector .x wherever it occurs, in a vectorized fashion.

unsalt(.x, .seed = NULL, .chars = letters, .n_chars = 5L)

Arguments

.x
a vector.

.seed
an integer to seed the random number generator.

.chars
set of characters to unsalt with.

.n_chars
an integer; number of characters to unsalt with.

Value

An unsalted version of the vector.

Details

The user is advised to check out Wikipedia for more information.

Examples

# Use various number of characters
unsalt(salt(letters, .n_chars = 0L))

#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z"

unsalt(salt(letters, .n_chars = 1L))

#>  [1] "faf" "fbf" "fcf" "fdf" "fef" "fff" "fgf" "fhf" "fif" "fjf" "fkf" "flf"
#> [13] "fmf" "fnf" "fof" "fpf" "fqf" "frf" "fsf" "ftf" "fuf" "fvf" "fwf" "fxf"
#> [25] "fyf" "fzf"

unsalt(salt(letters, .n_chars = 5L))

#>  [1] "jwimpajwimp" "jwimpbjwimp" "jwimpcjwimp" "jwimpdjwimp" "jwimpejwimp"
#>  [6] "jwimpfjwimp" "jwimpgjwimp" "jwimphjwimp" "jwimpijwimp" "jwimpjjwimp"
#> [11] "jwimpkjwimp" "jwimpljwimp" "jwimpmjwimp" "jwimpnjwimp" "jwimpojwimp"
#> [16] "jwimppjwimp" "jwimpqjwimp" "jwimprjwimp" "jwimpsjwimp" "jwimptjwimp"
#> [21] "jwimpujwimp" "jwimpvjwimp" "jwimpwjwimp" "jwimpxjwimp" "jwimpyjwimp"
#> [26] "jwimpzjwimp"

unsalt(salt(letters))

#>  [1] "cskvqacskvq" "cskvqbcskvq" "cskvqccskvq" "cskvqdcskvq" "cskvqecskvq"
#>  [6] "cskvqfcskvq" "cskvqgcskvq" "cskvqhcskvq" "cskvqicskvq" "cskvqjcskvq"
#> [11] "cskvqkcskvq" "cskvqlcskvq" "cskvqmcskvq" "cskvqncskvq" "cskvqocskvq"
#> [16] "cskvqpcskvq" "cskvqqcskvq" "cskvqrcskvq" "cskvqscskvq" "cskvqtcskvq"
#> [21] "cskvqucskvq" "cskvqvcskvq" "cskvqwcskvq" "cskvqxcskvq" "cskvqycskvq"
#> [26] "cskvqzcskvq"

# Use other sets of characters to salt with
unsalt(salt(letters, .chars = letters[1:2]), .chars = letters[1:2])

#>  [1] "abbabaabbab" "abbabbabbab" "abbabcabbab" "abbabdabbab" "abbabeabbab"
#>  [6] "abbabfabbab" "abbabgabbab" "abbabhabbab" "abbabiabbab" "abbabjabbab"
#> [11] "abbabkabbab" "abbablabbab" "abbabmabbab" "abbabnabbab" "abbaboabbab"
#> [16] "abbabpabbab" "abbabqabbab" "abbabrabbab" "abbabsabbab" "abbabtabbab"
#> [21] "abbabuabbab" "abbabvabbab" "abbabwabbab" "abbabxabbab" "abbabyabbab"
#> [26] "abbabzabbab"